requirements.txt
fastapi[all]
uvicorn
python-multipart
pillow
numpy
scikit-learn
xgboost
opencv-python-headless
joblib
tensorflow
